+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **Under development**
 
 - feat: vehicles and vehicle types are now always generated
+- feat: read vehicles data from zip files
 - feat : option parameter to remove filtering for requesting departements in hts
 - fix: secondary location model used same random seed in every parallel thread
 - feat: add a new method for attributing income to housholds using the bhepop2 package